agent-comms/requirements_standard_agent.txt
# Standard Agent Prompts - Requirements

# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Providers
openai>=1.10.0
# anthropic>=0.8.0  # Uncomment if using Anthropic
# google-generativeai>=0.3.0  # Uncomment if using Google Gemini

# Utilities
pandas>=2.0.0
matplotlib>=3.7.0
altair>=5.0.0

# Testing and development
pytest>=7.0.0